Change signature for Notify to take a custom type
This way we can pass the message, but also encapsulate some private
information and provide methods that can be called by subscribers. As an
example, we included the Reply method that can be used to post messages
to the chat, without the subscriber needing to know which channel the
producer is currently connected to.

The sample program also includes a subscriber with a bad joke about
saving.